Issue #2274: Move pinned tile package in channel package
