+++ app/src/test/java/org/mozilla/tv/firefox/channels/pinnedtile/PinnedTileTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.tv.firefox.pinnedtile
+package org.mozilla.tv.firefox.channels.pinnedtile
 
 import android.graphics.drawable.Drawable
 import androidx.lifecycle.MutableLiveData
